Linked bricks
One thing that bothered me while using LDR-Importer (super add-on by the way...) was that after import, you loose all references to the original bricks. Changing the lay-out of one type of brick is super-tedious if you have to repeat it for every instance of that brick.

Therefore...

I 'tweaked' the code a bit to automatically link identical bricks. It ignores all identical meshes and renames them to the form "brick name"+"colour". An extra checkbox has been added "Link Identical Bricks".

Note of warning: this is a 'tweak' and as such not the nicest way of implementing this feature. Just consider it a proof-of-concept. If you guys like it, I might do the effort to get it in a better shape. ;-)

Btw: This is my first experience on GitHub, so please forgive me if I make a boo boo...