Add logo function
The user now has the possibility to select an option that puts the LEGO logo on top off all studs in the file. A requirement for this is that the "logo3.dat" file from the LDraw parts tracker is located in the standard p folder.
Note that I also changed the threshold for the cleanup, because otherwise the logo would loose its geometry because it's so small. To make sure the thershold is fit for all import scales, the threshold is now multiplied by the import scale.